moving evictor to its own folder
